remove np.float

+++ src/correction_tools/bleedthrough.py
@@ -457,7 +457,7 @@ def Generate_bleedthrough_correction(bleed_folders,
         # calculate inverse matrix for each pixel
         if verbose:
             print(f"-- generating inverse matrix.")
-        _bleed_profiles = np.zeros(np.shape(bld_corr_profile), dtype=np.float)
+        _bleed_profiles = np.zeros(np.shape(bld_corr_profile), dtype=np.float32)
         for _i in range(np.shape(bld_corr_profile)[-2]):
             for _j in range(np.shape(bld_corr_profile)[-1]):
                 if generate_2d:
